теперь точно готово

+++ src/exercise5.c
@@ -1,21 +1,22 @@
 #include <stdio.h>
 
 int main(int argc, char *argv[]) {
-    int a[10][3];
-    int b[3][10];
-    int f[10][10];
+    double a[10][3]; 
+    double b[3][10];  
+    double f[10][10]; 
+
     for (int i = 0; i < 10; i++) {
         for (int j = 0; j < 3; j++) {
-            scanf("%d", &a[i][j]);
+            scanf("%lf", &a[i][j]); 
         }
     }
     
     for (int i = 0; i < 3; i++) {
         for (int j = 0; j < 10; j++) {
-            scanf("%d", &b[i][j]);
+            scanf("%lf", &b[i][j]); 
         }
     }
-    
+
     for (int i = 0; i < 10; i++) {
         for (int j = 0; j < 10; j++) {
             f[i][j] = 0;
@@ -24,14 +25,14 @@ int main(int argc, char *argv[]) {
             }
         }
     }
-
+    
     for (int i = 0; i < 10; i++) {
         for (int j = 0; j < 10; j++) {
-            printf("%d", f[i][j]);
+            printf("%.2f", f[i][j]);  
             if (i < 9 || j < 9) printf(" ");
         }
     }
     printf("\n");
     
     return 0;
-}
+}
